refactor(scripts): extract shared story-reference resolver
- New scripts/lib/story-refs.mjs resolves a PR/CLI reference to an existing story slug (exact slug, story-file path, or "Story X.Y")
- mark-story-done.mjs now imports it instead of its own copy, so the automation and the upcoming PR check stay consistent (Story 11.7)
